Corrigindo erro da conta 5

+++ projeto_thrust.cu
@@ -300,9 +300,9 @@ int main(){
 
                 //swap(vec_pontos[count], vec_pontos[count+1]);
 
-                thrust::counting_iterator<float> iterador(0);
+                thrust::counting_iterator<int> iterador(0);
 
-                thrust::transform((float)iterador, (float)iterador+(n-1), array_float.begin(), swap_vec(thrust::raw_pointer_cast(array.data())));
+                thrust::transform(iterador, iterador+(n-1), array_float.begin(), swap_vec(thrust::raw_pointer_cast(array.data())));
                 for(int it = 0; it < n; it++){
                   std::cout << "swap vec_pontos: " << static_cast<graph>(vec_pontos[it]).id << std::endl;
                 }
